changed uint32_t len parm in strndup to size_t

+++ ft_strndup.c
@@ -6,13 +6,13 @@
 /*   By: sclolus <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/03/18 04:58:30 by sclolus           #+#    #+#             */
-/*   Updated: 2017/03/18 05:00:08 by sclolus          ###   ########.fr       */
+/*   Updated: 2017/09/12 07:16:13 by sclolus          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-char	*ft_strndup(char *str, uint32_t len)
+char	*ft_strndup(char *str, size_t len)
 {
 	char	*tmp;
